Cast field to long (#1716)

+++ common/persistence/src/main/java/app/coronawarn/server/common/persistence/service/common/KeySharingPoliciesChecker.java
@@ -85,7 +85,7 @@ public class KeySharingPoliciesChecker {
   private LocalDateTime getRollingPeriodExpiryTime(DiagnosisKey diagnosisKey) {
     return LocalDateTime
         .ofEpochSecond(diagnosisKey.getRollingStartIntervalNumber() * TEN_MINUTES_INTERVAL_SECONDS, 0, UTC)
-        .plusMinutes(diagnosisKey.getRollingPeriod() * DiagnosisKey.ROLLING_PERIOD_MINUTES_INTERVAL);
+        .plusMinutes((long) diagnosisKey.getRollingPeriod() * DiagnosisKey.ROLLING_PERIOD_MINUTES_INTERVAL);
   }
 
   /**
